Minor Code Cleanup
i saw it, and it made my eyes hurt. I guess sometimes seeing different solutions or possibilities, well not sometimes but most times, it's all reliant on your headspace.
